backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
python-multipart>=0.0.9
aiofiles>=23.0.0
Pillow>=10.0.0
onnxruntime>=1.18.0
huggingface-hub>=0.23.0
pandas>=2.0.0
numpy>=1.26.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
